New C implementation of the Masking
This commit proposes a new C implementation of the masking for the
linear trails. Masks are generated directly in Python to a C file.
This prevents bugs such as 1 being considered uint32_t instead of
uint64_t... The selection of the test is done at build time.
By default the full trail on mini_morus_640 will be executed.
To build the full trail on mini_morus_1280 the target is full_64
Other targets are available and presented by the default make.